+++ src/js/MouseHandler.d.ts
@@ -85,7 +85,7 @@ export declare class MouseHandler {
      *          .click( (e:XMouseEvent) => {
      *              console.log( 'Click.' );
      *          } )
-     *          .wheel( (e:XMouseEvent) => {
+     *          .wheel( (e:XWheelEvent) => {
      *              console.log( 'Wheel. delta='+e.deltaY );
      *          } )
      *