+++ src/ts/MouseHandler.ts
@@ -91,7 +91,7 @@ export class MouseHandler {
      *          .click( (e:XMouseEvent) => {
      *              console.log( 'Click.' );
      *          } )
-     *          .wheel( (e:XMouseEvent) => {
+     *          .wheel( (e:XWheelEvent) => {
      *              console.log( 'Wheel. delta='+e.deltaY );
      *          } )
      *