+++ src/ts/Polygon.ts
@@ -14,8 +14,9 @@
  * @modified 2019-11-22 Added the rotate(number,Vertex) function.
  * @modified 2020-03-24 Ported this class from vanilla-JS to Typescript.
  * @modified 2020-10-30 Added the `addVertex` function.
- * @modified 2020-10-31 Added thet `getVertexAt` function.
- * @version 1.3.0
+ * @modified 2020-10-31 Added the `getVertexAt` function.
+ * @modified 2020-11-06 Added the `move` function.  
+ * @version 1.4.0
  *
  * @file Polygon
  * @public
@@ -98,6 +99,23 @@ export class Polygon implements SVGSerializable {
 	    return this.vertices[ this.vertices.length - (Math.abs(index)%this.vertices.length) ];
 	else
 	    return this.vertices[ index % this.vertices.length ];
+    };
+
+
+    /**
+     * Move the polygon's vertices by the given amount.
+     *
+     * @method move
+     * @param {XYCoords} amount - The amount to move.
+     * @instance
+     * @memberof Polygon
+     * @return {Polygon} this for chaining
+     **/
+    move( vert:Vertex ) : Polygon {
+	for( var i in this.vertices ) {
+	    this.vertices[i].add( vert );
+	}
+	return this;
     };