+++ src/ts/utils/algorithms/convexHull.ts
@@ -48,6 +48,7 @@ export const getConvexHull = (points:Array<XYCoords>) : Array<XYCoords> => {
  * Compare two vertices to create an order on the 2D plane.
  *
  * @name comparison
+ * @private
  * @param {XYCoords} a - The first of the two points to compare.
  * @param {XYCoords} b - The second of the two points to compare.
  * @return {number} A number indicating the order (negative if `a` is 'smaller', 0 if both are equal, positive if `a` is 'larger').
@@ -60,6 +61,8 @@ const comparison = (a:XYCoords, b:XYCoords) : number => {
 /**
  * Calculate the cross product of the three coordinates, interpreted as vectors.
  * 
+ * @name cross
+ * @private
  * @param {XYCoords} a
  * @param {XYCoords} b
  * @param {XYCoords} o