+++ docs_jekyll/changelog.md
@@ -1,10 +1,13 @@
 ---
 layout: page
+title: Changelog
 ---
 
 
 # Changelog
-* 2020-11-01
+* 2020-10-14
+  * Changed `window` to `globalThis`.
+* 2020-10-01
   * Migrated the build tools from Webpack-4 to Webpack-5.
   * Moved the build scripts to ./bin.
 * 2020-10-31